+++ CHANGELOG.md
@@ -4,13 +4,13 @@ All notable changes to this project will be documented in this file.
 ## UNRELEASED
 ### Added
 * ui/ui-toolkit: i18n using the Lokalise
+* Autofill support for PAYONE credit card forms
 * core/ui: Add support for project specific custom properties
   * Make the external billing subject text max length configurable
 ### Changed
 ### Removed
 * ui: Remove phrase for i18n
 ### Fixed
-
 
 ## [0.72.5]
 ### Changed